requirements.txt
python-dotenv
langchain 
langchain-google-genai 
langchain-groq
google-ai-generativelanguage==0.6.15
